requirements-dev.txt
# Testing dependencies
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
responses==0.25.0
python-dotenv==1.0.0

# Production dependencies
-r requirements.txt
